% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAR.R
\name{MAR}
\alias{MAR}
\title{Doubly robust inference with non-probability survey samples}
\usage{
MAR(SA, SB)
}
\arguments{
\item{SA}{A matrix, the non-probability sample. The last column indicates the outcome variable and the other columns represent the covariates.}

\item{SB}{A matrix, the reference probability sample. The last column indicates the survey weights and the other columns represent the covariates.}
}
\value{
A list with two elements:
\itemize{
  \item mu, the IPW, regression, and double robust estimates of the population mean.
  \item theta, the maximum pseudo-likelihood estimates of parameters in the participation probability model.
}
}
\description{
Function to apply the pseudo-likelihood method in the assumption of a ignorable participation mechanism.
}
\references{
Chen, Y., Li, P. and Wu, C. (2020).
Doubly robust inference with nonprobability survey samples.
\emph{Journal of the American Statistical Association} \strong{115}, 2011–2021.
}
