% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNAR.R
\name{MNAR}
\alias{MNAR}
\title{Doubly robust inference with nonprobability survey samples}
\usage{
MNAR(SA, SB, par0 = rep(0, ncol(SA)))
}
\arguments{
\item{SA}{A matrix, the non-probability  sample. The last column indicates the outcome variable and the other columns represent the covariates. Additionally, the second-to-last column serves as the instrumental variable.}

\item{SB}{A matrix, the reference probability sample. The last column indicates the survey weights and the other columns represent the covariates. Additionally, the second-to-last column serves as the instrumental variable.}

\item{par0}{A vector, the initial value of the parameter vector in the participation probability model. The default value is set to zero.}
}
\value{
A list with three elements:
\itemize{
  \item theta, the maximum pseudo-likelihood estimates of the parameters in the participation probability model.
  \item xi, the maximum likelihood estimates of the parameters in the outcome regression model.
  \item mu, the regression, IPW, and AIPW estimates of the population mean.
}
}
\description{
Function to apply the pseudo-likelihood method in the assumption of a nonignorable participation mechanism.
}
\references{
Liu, Y., Yuan, M., Li, P. and Wu, C. (2024).
Statistical inference with nonignorable non-probability survey samples.
\emph{arXiv}: 2410.02920.
}
