% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seMNAR.R
\name{seMNAR}
\alias{seMNAR}
\title{Doubly robust inference with nonprobability survey samples}
\usage{
seMNAR(SA, SB, estMNAR, secOrderProb = "Hajek", N0 = NULL)
}
\arguments{
\item{SA}{A matrix, the non-probability  sample. The last column indicates the outcome variable and the other columns represent the covariates. Additionally, the second-to-last column serves as the instrumental variable.}

\item{SB}{A matrix, the reference probability sample. The last column indicates the survey weights and the other columns represent the covariates. Additionally, the second-to-last column serves as the instrumental variable.}

\item{estMNAR}{A list, the output object of the function *NMAR()*.}

\item{secOrderProb}{A character. "Hajek" indicates that the second-order selection probability is estimated using the Hajek (1964)'s approximation. Otherwise, it is estimated through simple random sampling without replacement.}

\item{N0}{A numeric, the size of the finite population. If *NULL*, it is set to the sum of weights from the reference probability sample.}
}
\value{
A list with three elements:
\itemize{
  \item seTheta, the standard errors of the maximum pseudo-likelihood estimates of the parameters in the participation probability model.
  \item seXi, the standard errors of the maximum likelihood estimates of the parameters in the outcome regression model.
  \item seMu, the standard errors of the regression, IPW, and AIPW estimates of the population mean.
}
}
\description{
Function to apply the pseudo-likelihood method in the assumption of a nonignorable participation mechanism.
}
\references{
Hajek, J. (1964). Asymptotic theory of rejective sampling with varying probabilities from a finite population. \emph{Annals of Mathematical Statistics} **35**, 1491–1523.

Liu, Y., Yuan, M., Li, P. and Wu, C. (2024).
Statistical inference with nonignorable non-probability survey samples.
\emph{arXiv}: 2410.02920.
}
